+++ app/src/main/java/com/example/movieapp/activity/MainActivity.kt
@@ -1,7 +1,8 @@
-package com.example.movieapp
+package com.example.movieapp.activity
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.example.movieapp.R
 import com.example.movieapp.ui.main.MainFragment
 
 class MainActivity : AppCompatActivity() {
