+++ app/src/main/java/com/example/movieapp/api/MovieApi.kt
@@ -1,5 +1,6 @@
-package com.example.movieapp
+package com.example.movieapp.api
 
+import com.example.movieapp.model.Movie
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Query
