+++ app/src/main/java/com/example/movieapp/model/Movie.kt
@@ -1,4 +1,4 @@
-package com.example.movieapp
+package com.example.movieapp.model
 
 import com.google.gson.annotations.SerializedName
